# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: "Final API"
  description: "Final API"
  version: "1.0.0"
servers:
  - url: "https://localhost:6060"
paths:
...
  /api/auth/signin:
    post:
      summary: "Authenticate user"
      operationId: "authUser"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        "200":
          description: "OK"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JwtResponse'
        "400":
          description: "Bad Request"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
components:
  schemas:
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
    Profile:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
    JwtResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        is_active:
          type: boolean
        is_superuser:
          type: boolean
        last_login:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
        profile:
          $ref: '#/components/schemas/Profile'
        access_token:
          type: string
        refresh_token:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...